[major] try in vain to get it to compile against 1.7.10
