+++ ballast-core/build.gradle.kts
@@ -1,6 +1,6 @@
 plugins {
-    id("com.android.library")
-    kotlin("multiplatform")
+    `copper-leaf-android`
+    `copper-leaf-targets`
     `copper-leaf-base`
     `copper-leaf-version`
     `copper-leaf-lint`
@@ -9,68 +9,8 @@ plugins {
 
 description = "Opinionated Application State Management framework for Kotlin Multiplatform"
 
-android {
-    compileSdk = 31
-    defaultConfig {
-        minSdk = 21
-        targetSdk = 31
-        testInstrumentationRunner = "android.support.test.runner.AndroidJUnitRunner"
-    }
-    buildTypes {
-        val release by getting {
-            isMinifyEnabled = false
-        }
-    }
-    sourceSets {
-        getByName("main") {
-            setRoot("src/androidMain")
-        }
-        getByName("androidTest") {
-            setRoot("src/androidTest")
-        }
-    }
-    testOptions {
-        unitTests {
-            isIncludeAndroidResources = true
-            isReturnDefaultValues = true
-        }
-    }
-    lint {
-        disable("GradleDependency")
-    }
-}
-
 kotlin {
-    explicitApi()
-
-    // targets
-    jvm { }
-    android {
-        publishAllLibraryVariants()
-    }
-    js(BOTH) {
-        browser {
-            testTask {
-                enabled = false
-            }
-        }
-    }
-    nativeTargetGroup(
-        "ios",
-        iosArm32(),
-        iosArm64(),
-        iosX64(),
-        iosSimulatorArm64(),
-    )
-
-    // sourcesets
     sourceSets {
-        all {
-            languageSettings.apply {
-            }
-        }
-
-        // Common Sourcesets
         val commonMain by getting {
             dependencies {
                 api(libs.kotlinx.coroutines.core)
@@ -78,24 +18,13 @@ kotlin {
         }
         val commonTest by getting {
             dependencies {
-                implementation(kotlin("test-common"))
-                implementation(kotlin("test-annotations-common"))
                 implementation(project(":ballast-test"))
             }
         }
-
         val jvmMain by getting {
             dependencies {
             }
         }
-        val jvmTest by getting {
-            dependencies {
-                implementation(kotlin("test-junit"))
-                implementation(libs.kotlinx.coroutines.test)
-            }
-        }
-
-        // Android JVM Sourcesets
         val androidMain by getting {
             dependencies {
                 api(libs.androidx.lifecycle.common)
@@ -103,45 +32,11 @@ kotlin {
                 api(libs.androidx.lifecycle.viewmodel)
             }
         }
-        val androidAndroidTestRelease by getting { }
-        val androidTest by getting {
-            dependsOn(androidAndroidTestRelease)
-            dependencies {
-                implementation(kotlin("test-junit"))
-            }
-        }
-
-        // JS Sourcesets
         val jsMain by getting {
             dependencies { }
         }
-        val jsTest by getting {
-            dependencies {
-                implementation(kotlin("test-js"))
-            }
-        }
-
-        // iOS Sourcesets
         val iosMain by getting {
             dependencies { }
         }
-        val iosTest by getting {
-            dependencies { }
-        }
-    }
-}
-
-tasks.withType<JavaCompile> {
-    sourceCompatibility = Config.javaVersion
-    targetCompatibility = Config.javaVersion
-}
-tasks.withType<Test> {
-    testLogging {
-        showStandardStreams = true
-    }
-}
-tasks.withType<org.jetbrains.kotlin.gradle.tasks.KotlinCompile> {
-    kotlinOptions {
-        jvmTarget = Config.javaVersion
     }
 }
