+++ ballast-firebase-crashlytics/build.gradle.kts
@@ -1,116 +1,35 @@
 plugins {
-    id("com.android.library")
-    kotlin("multiplatform")
+    `copper-leaf-android`
+    `copper-leaf-targets`
     `copper-leaf-base`
     `copper-leaf-version`
     `copper-leaf-lint`
     `copper-leaf-publish`
 }
 
-description = "Opinionated Application State Management framework for Kotlin Multiplatform"
-
-android {
-    compileSdkVersion(31)
-    defaultConfig {
-        minSdkVersion(21)
-        targetSdkVersion(31)
-        testInstrumentationRunner = "android.support.test.runner.AndroidJUnitRunner"
-    }
-    buildTypes {
-        val release by getting {
-            isMinifyEnabled = false
-        }
-    }
-    sourceSets {
-        getByName("main") {
-            setRoot("src/androidMain")
-        }
-        getByName("androidTest") {
-            setRoot("src/androidTest")
-        }
-    }
-    testOptions {
-        unitTests {
-            isIncludeAndroidResources = true
-            isReturnDefaultValues = true
-        }
-    }
-    lintOptions {
-        disable("GradleDependency")
-    }
-}
+description = "Opinionated Kotlin multiplatform Repository Caching library based on Ballast MVI"
 
 kotlin {
-    explicitApi()
-
-    // targets
-    jvm { }
-    android {
-        publishAllLibraryVariants()
-    }
-    js(BOTH) {
-        browser {
-            testTask {
-                enabled = false
-            }
-        }
-    }
-    nativeTargetGroup(
-        "ios",
-        iosArm32(),
-        iosArm64(),
-        iosX64(),
-        iosSimulatorArm64(),
-    )
-
-    // sourcesets
     sourceSets {
-        all {
-            languageSettings.apply {
-            }
-        }
-
-        // Common Sourcesets
         val commonMain by getting {
             dependencies {
+                implementation(project(":ballast-core"))
             }
         }
-        val commonTest by getting {
-            dependencies {
-                implementation(kotlin("test-common"))
-                implementation(kotlin("test-annotations-common"))
-            }
+        val jvmMain by getting {
+            dependencies { }
         }
-
-        // Android JVM Sourcesets
         val androidMain by getting {
             dependencies {
-                api(project(":ballast-core"))
                 implementation(project.dependencies.platform(libs.firebase.bom))
                 implementation(libs.firebase.crashlytics)
             }
         }
-        val androidAndroidTestRelease by getting { }
-        val androidTest by getting {
-            dependsOn(androidAndroidTestRelease)
-            dependencies {
-                implementation(kotlin("test-junit"))
-            }
+        val jsMain by getting {
+            dependencies { }
         }
-    }
-}
-
-tasks.withType<JavaCompile> {
-    sourceCompatibility = Config.javaVersion
-    targetCompatibility = Config.javaVersion
-}
-tasks.withType<Test> {
-    testLogging {
-        showStandardStreams = true
-    }
-}
-tasks.withType<org.jetbrains.kotlin.gradle.tasks.KotlinCompile> {
-    kotlinOptions {
-        jvmTarget = Config.javaVersion
+        val iosMain by getting {
+            dependencies { }
+        }
     }
 }
