+++ build.gradle.kts
@@ -1,7 +1,7 @@
 plugins {
     `copper-leaf-base`
     `copper-leaf-version`
-    id("org.jetbrains.kotlinx.binary-compatibility-validator") version "0.9.0"
+    id("org.jetbrains.kotlinx.binary-compatibility-validator") version "0.11.0"
     `copper-leaf-sonatype`
 }
