+++ buildSrc/src/main/kotlin/copper-leaf-android.gradle.kts
@@ -1,0 +1,59 @@
+@file:Suppress("UnstableApiUsage", "UNUSED_VARIABLE")
+
+import org.jetbrains.kotlin.gradle.dsl.KotlinMultiplatformExtension
+
+plugins {
+    id("com.android.library")
+}
+
+android {
+    compileSdk = 31
+    defaultConfig {
+        minSdk = 21
+        targetSdk = 31
+        testInstrumentationRunner = "android.support.test.runner.AndroidJUnitRunner"
+    }
+    buildTypes {
+        val release by getting {
+            isMinifyEnabled = false
+        }
+    }
+    sourceSets {
+        getByName("main") {
+            setRoot("src/androidMain")
+        }
+        getByName("androidTest") {
+            setRoot("src/androidTest")
+        }
+    }
+    testOptions {
+        unitTests {
+            isIncludeAndroidResources = true
+            isReturnDefaultValues = true
+        }
+    }
+    lint {
+        disable += listOf("GradleDependency")
+    }
+    publishing {
+        multipleVariants {
+            withSourcesJar()
+            withJavadocJar()
+            allVariants()
+        }
+    }
+}
+
+afterEvaluate {
+    // Remove log pollution until Android support in KMP improves.
+    project.extensions.findByType<KotlinMultiplatformExtension>()?.let { kmpExt ->
+        kmpExt.sourceSets.removeAll {
+            setOf(
+                "androidAndroidTestRelease",
+                "androidTestFixtures",
+                "androidTestFixturesDebug",
+                "androidTestFixturesRelease",
+            ).contains(it.name)
+        }
+    }
+}
