+++ buildSrc/src/main/kotlin/copper-leaf-targets.gradle.kts
@@ -1,0 +1,104 @@
+@file:Suppress("UNUSED_VARIABLE")
+
+plugins {
+    kotlin("multiplatform")
+}
+
+kotlin {
+    explicitApi()
+
+    // targets
+    jvm { }
+    android {
+        publishAllLibraryVariants()
+    }
+    js(BOTH) {
+        browser {
+            testTask {
+                enabled = false
+            }
+        }
+    }
+    nativeTargetGroup(
+        "ios",
+        iosArm32(),
+        iosArm64(),
+        iosX64(),
+        iosSimulatorArm64(),
+    )
+
+    // sourcesets
+    sourceSets {
+        all {
+            languageSettings.apply {
+            }
+        }
+
+        // Common Sourcesets
+        val commonMain by getting {
+            dependencies {
+            }
+        }
+        val commonTest by getting {
+            dependencies {
+                implementation(kotlin("test-common"))
+                implementation(kotlin("test-annotations-common"))
+            }
+        }
+
+        val jvmMain by getting {
+            dependencies {
+            }
+        }
+        val jvmTest by getting {
+            dependencies {
+                implementation(kotlin("test-junit"))
+                implementation("org.jetbrains.kotlinx:kotlinx-coroutines-test:1.6.4")
+            }
+        }
+
+        // Android JVM Sourcesets
+        val androidMain by getting {
+            dependencies {
+            }
+        }
+        val androidTest by getting {
+            dependencies {
+                implementation(kotlin("test-junit"))
+            }
+        }
+
+        // JS Sourcesets
+        val jsMain by getting {
+            dependencies { }
+        }
+        val jsTest by getting {
+            dependencies {
+                implementation(kotlin("test-js"))
+            }
+        }
+
+        // iOS Sourcesets
+        val iosMain by getting {
+            dependencies { }
+        }
+        val iosTest by getting {
+            dependencies { }
+        }
+    }
+}
+
+tasks.withType<JavaCompile> {
+    sourceCompatibility = Config.javaVersion
+    targetCompatibility = Config.javaVersion
+}
+tasks.withType<Test> {
+    testLogging {
+        showStandardStreams = true
+    }
+}
+tasks.withType<org.jetbrains.kotlin.gradle.tasks.KotlinCompile> {
+    kotlinOptions {
+        jvmTarget = Config.javaVersion
+    }
+}
