+++ gradle.properties
@@ -12,8 +12,5 @@ android.useAndroidX=true
 
 #MPP
 kotlin.mpp.stability.nowarn=true
-kotlin.mpp.enableGranularSourceSetsMetadata=true
-kotlin.mpp.enableCompatibilityMetadataVariant=true
-kotlin.native.enableDependencyPropagation=false
 kotlin.mpp.enableCInteropCommonization=true
 kotlin.native.binary.memoryModel=experimental
